Rework on CG and length
