+++ src/fastga_he/models/propulsion/components/connectors/h2_fuel_system/constants.py
@@ -2,14 +2,4 @@
 # Electric Aircraft.
 # Copyright (C) 2025 ISAE-SUPAERO
 
-POSSIBLE_POSITION = [
-    "from_rear_to_front",
-    "from_rear_to_center",
-    "from_rear_to_wing",
-    "from_center_to_wing",
-    "from_center_to_front",
-    "from_front_to_wing",
-    "in_the_back",
-    "at_center",
-    "in_the_wing",
-]
+POSSIBLE_POSITION = ["in_the_front", "in_the_middle", "in_the_rear"]
